requirements.txt
streamlit
langchain
openai
fastapi
uvicorn
langchain-community
langchain-core
flask
google-generativeai